+++ lib/src/eventargs.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an Apache-2.0 license that can be
 // found in the LICENSE file.
 
-import 'package:event/event.dart';
+import 'event.dart';
 
 /// An abstract representation of the (optional) arguments provided
 /// to handlers when an [Event] occurs.
@@ -10,45 +10,44 @@ import 'package:event/event.dart';
 /// This is intended to be extended with your own custom
 /// class, if you want to provide subscribers with some
 /// data relating to the [Event].  Alternatively, use one
-/// of the supplied [EventArgs] derived classes: [StdEventArgs],
-/// [Value] or [Values].
+/// of the three supplied [EventArgs] derived classes:
+/// [WhenWhy], [Value] or [Values].
 ///
 /// ```dart
 /// // example
-/// class ValueEventArgs extends EventArgs {
+/// class MyChangedValue extends EventArgs {
 ///   int changedValue;
-///   ValueEventArgs(this.changedValue);
+///   MyChangedValue(this.changedValue);
 /// }
 /// ```
 abstract class EventArgs {}
 
-/// An [EventArgs] derived class that includes the date and
-/// time the [Event] was broadcast and an
-/// optional description.
+/// An [EventArgs] derived class that includes the date/time
+/// the [Event] was broadcast, and an optional description.
 ///
 /// Provides a [whenOccurred] and [description] fields.
 ///
 /// ```dart
 /// // example declaration ...
-/// var myEvent = Event<StdEventArgs>();
+/// var myEvent = Event<WhenWhy>();
 /// // broadcast with an optional description ...
-/// myEvent.broadcast(StdEventArgs(description: 'something'))
+/// myEvent.broadcast(WhenWhy(description: 'something'))
 /// // in subscriber handler ...
 /// print(args.whenOccurred);
 /// print(args.description);
 /// ```
-class StdEventArgs extends EventArgs {
+class WhenWhy extends EventArgs {
   /// The date and time the [Event] was broadcast.
   DateTime whenOccurred;
 
   /// An optional description or other information.
   String description;
 
-  /// Creates a new [StdEventArgs], with an optional description.
-  StdEventArgs({this.description = ''}) : whenOccurred = DateTime.now();
+  /// Creates a new [WhenWhy], with an optional description.
+  WhenWhy({this.description = ''}) : whenOccurred = DateTime.now();
 }
 
-/// Represents a [StdEventArgs] derived class with one (generic) value.
+/// Represents an [EventArgs] derived class with one (generic) value.
 ///
 /// For use as a quick alternative to defining your own custom EventArgs
 /// class. Provides a [value] field that contains the value supplied on
@@ -56,24 +55,25 @@ class StdEventArgs extends EventArgs {
 ///
 /// See also [Values] which supports 2 (generic) values.
 ///
+///
 /// ```dart
 /// // example declaration with an inferred type
-/// var e = Event<GenericEventArgs1>();
-/// // example declaration with an declared type
-/// var e = Event<GenericEventArgs1<String>>();
+/// var e = Event<Value>();
+/// // equivalent example declaration with an declared type
+/// var e = Event<Value<String>>();
 /// e.subscribe((args) => print(args.value));
-/// e.broadcast(GenericEventArgs1('hello'));
+/// e.broadcast(Value('hello'));
 /// // outputs: hello
 ///```
-class Value<T> extends StdEventArgs {
+class Value<T> extends EventArgs {
   /// A generic value.
   T value;
 
   /// Creates a [Value] representing one generic [value]
-  Value(this.value, {String description = ''}) : super(description: description);
+  Value(this.value);
 }
 
-/// Represents a [StdEventArgs] derived class with two (generic) values.
+/// Represents an [EventArgs] derived class with two (generic) values.
 ///
 /// For use as a quick alternative to defining your own custom EventArgs
 /// class. Provides [value1] and [value2] fields containing the values
@@ -83,28 +83,16 @@ class Value<T> extends StdEventArgs {
 ///
 /// ```dart
 /// // example
-/// var e = Event<EventArgs2<String, int>>();
+/// var e = Event<Values<String, int>>();
 /// e.subscribe((args) => print('${args.value1} : ${args.value2}'));
-/// e.broadcast(EventArgs2('boom', 37));
+/// e.broadcast(Values('boom', 37));
 /// // outputs: boom : 37
 ///```
-class Values<T1, T2> extends StdEventArgs {
+class Values<T1, T2> extends EventArgs {
   /// A generic value.
   T1 value1;
   T2 value2;
 
   /// Creates an [Values] representing two generic values: [value1] and [value2]
-  Values(this.value1, this.value2, {String description = ''}) : super(description: description);
+  Values(this.value1, this.value2);
 }
-
-/// Represents an empty [EventArg] derived class. For use where no
-/// arguments are required, and one wants to make explicit the generic
-/// type.
-/// ```dart
-/// var e = Event<EmptyEventArgs>();
-/// // is equivalent to...
-/// var e = Event();
-/// ```
-/// The difference is the first above makes it more explicit to a code
-/// reader that there are no EventArgs.
-class EmptyEventArgs extends EventArgs {}